major refactor including directory structure and imports. most importantly this should work with systemd timers now
