+++ waveshare_lib/epd7in5.py
