+++ waveshare_lib/epdconfig.py
